+++ packages/components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @hashicorp/design-system-components
 
+## 0.12.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @hashicorp/ember-flight-icons@2.0.7
+
+## 0.12.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @hashicorp/ember-flight-icons@2.0.6
+
 ## 0.12.0
 
 ### Minor Changes